Create version 0.0.1 of the Odin Calculator
Publish the code for the first working version of the Calculator for the
Odin Project. The Calculator can currently perform the basic arithmetic
by a user inputing a number, operand and number through some buttons,
and a result being displayed as the innerText of an h1 element. In the
next version, I'll be looking into being able to show the previous
equation and being able to handle decimal numbers as inputs. Styling of
the Calculator is bare bones right now, I want to focus on the code.